Relocate a few classes to a more Hexagonal architecture style structure
