--- src/main/java/disney/reservation/notification/Adapter/Logger/Logger.java
@@ -1,5 +1,0 @@
-package disney.reservation.notification.Adapter.Logger;
-
-public interface Logger {
-    public void info(String message);
-}
