+++ src/main/java/disney/reservation/notification/domain/log/Logger.java
@@ -1,0 +1,5 @@
+package disney.reservation.notification.domain.log;
+
+public interface Logger {
+    void info(String message);
+}
